Restrucutre project folders
