+++ components/Footer/footer.css
@@ -1,0 +1,34 @@
+footer {
+    width: 100%;
+    height: 70px;
+    margin-top: 20px;
+}
+footer .footer-wrapper {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    font-size: 1.1rem;
+}
+footer .footer-wrapper .rss-content {
+    width: 25%;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    align-items: center;
+}
+footer .developedBy {
+    margin-top: 10px;
+    padding-bottom: 10px;
+}
+@media only screen and (max-width: 600px) {
+    footer .rss-content {
+        width: 100% !important;
+        display: flex;
+        flex-direction: row;
+        justify-content: space-around;
+        padding: 0px 20px;
+    }
+}
