+++ components/Separator/Separator.js
