+++ components/ToggleTheme/ToggleThemeColorButton.js
@@ -1,6 +1,6 @@
 import { Button, useColorMode } from "@chakra-ui/react"
-import NightIcon from "./icons/NightIcon"
-import SunIcon from "./icons/SunIcon"
+import NightIcon from "../icons/NightIcon"
+import SunIcon from "../icons/SunIcon"
 
 const ToggleThemeColorButton = () => {
     const { colorMode, toggleColorMode } = useColorMode()
