+++ components/sections/HomeSection/home.css
@@ -1,0 +1,71 @@
+.homeSectionContent {
+    width: 100%;
+    height: auto;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    position: relative;
+}
+
+.homeImageContent {
+    width: 40%;
+    height: auto;
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-end;
+    padding: 10px 25px 10px 15px;
+}
+
+.homeImageContent img {
+    width: 50%;
+    height: auto;
+    border-radius: 100%;
+    object-position: right;
+}
+
+.homeTitleContent {
+    width: 60%;
+    height: auto;
+}
+
+.homeText {
+    width: 75%;
+}
+
+.lastHomeTitle {
+    margin-bottom: 22px;
+}
+
+@media only screen and (max-width: 600px) {
+    .homeSectionContent {
+        margin-top: 60px;
+    }
+
+    .homeTitleContent {
+        width: 90%;
+        height: auto;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+    }
+
+    .homeTitle {
+        font-size: 34px !important;
+        text-align: center;
+    }
+
+    .lastHomeTitle {
+        margin-bottom: 24px;
+    }
+
+    .homeTitleContent {
+        width: 100%;
+    }
+
+    .homeImageContent {
+        display: none;
+    }
+}
